Extract uint_to_felt from u128, and added it for u8.
commit-id:ea76de16
